# expression: implement vectorized evaluation for `builtinCastStringAsIntSig`