Signed-off-by: AerysNan <aerysnan@gmail.com>

PCP #12103

### What problem does this PR solve? <!--add issue link with summary if exists-->

Implement vectorized evaluation for builtinTidbParseTsoSig.

### What is changed and how it works?

```
BenchmarkVectorizedBuiltinCastFunc/builtinCastStringAsIntSig-VecBuiltinFunc-4               7000            154545 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinCastFunc/builtinCastStringAsIntSig-NonVecBuiltinFunc-4            7238            167614 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinCastFunc/builtinCastStringAsIntSig-VecBuiltinFunc#01-4            8276            138306 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinCastFunc/builtinCastStringAsIntSig-NonVecBuiltinFunc#01-4         7501            161245 ns/op               0 B/op          0 allocs/op   
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test